First Step
The core of having a proper model in machine learning, based on what I have understood from the book Mathematics for Machine Learning by Marc Peter Deisenroth, A. Aldo Faisal, and Cheng Soon Ong, is finding the best parameters for our predictor. In this case, we are using Linear Regression.

For educational purposes, I have implemented an Empirical Risk Minimization (ERM) algorithm that is not optimized, with the goal of better understanding the mathematics behind machine learning.

The way I am searching for these parameters is abrupt and slow, but I want to see how the data responds. After this, I will try to optimize the algorithm.